@@ -1,6 +1,6 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
     <RootNamespace>ReactDashboardApp</RootNamespace>
   </PropertyGroup>
   <ItemGroup>
@@ -8,9 +8,10 @@
     <Folder Include="wwwroot\dist\" />
   </ItemGroup>
   <ItemGroup>
-    <PackageReference Include="DotNetify.Elements" Version="1.0.0" />
-    <PackageReference Include="DotNetify.SignalR" Version="3.4.0" />
-    <PackageReference Include="Microsoft.AspNetCore.App" />
+    <PackageReference Include="DotNetify.Elements" Version="1.3.0" />
+    <PackageReference Include="DotNetify.SignalR" Version="4.0.0" />
+    <PackageReference Include="Microsoft.AspNetCore.SpaServices" Version="3.0.0" />
+    <PackageReference Include="Microsoft.Extensions.Http" Version="3.1.4" />
     <PackageReference Include="System.Reactive" Version="3.1.1" />
     <PackageReference Include="Bogus" Version="22.1.3" />    
   </ItemGroup>
