@@ -5,6 +5,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.SpaServices.Webpack;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 using Shared;
 using System.Collections.Generic;
 using System.IO;
@@ -28,6 +29,8 @@ namespace Portal
          services.AddOptions();
          services.Configure<IdentityServerSettings>(identityServerSettings);
 
+         services.AddControllers();
+
          services.AddIdentityServer()
            .AddDeveloperSigningCredential()
            .AddInMemoryApiResources(IdentityServerConfig.GetApiResources())
@@ -41,33 +44,35 @@ namespace Portal
                options.ApiName = IdentityServerConfig.MyClientId;
             });
 
-         services.AddAuthorization()
-           .AddMvcCore()
-           .AddJsonFormatters();
+         services.AddAuthorization();
 
          services.AddHttpClient();
          services.AddResponseCompression();
       }
 
-      public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+      public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
       {
-         if(env.IsDevelopment())
+         if (env.IsDevelopment())
          {
+#pragma warning disable 618
             app.UseWebpackDevMiddleware(new WebpackDevMiddlewareOptions
             {
                HotModuleReplacement = true,
                HotModuleReplacementClientOptions = new Dictionary<string, string> { { "reload", "true" } },
             });
+#pragma warning restore 618
          }
 
          app.UseIdentityServer();
 
+         app.UseRouting();
+         app.UseEndpoints(endpoints => endpoints.MapControllers());
+
          app.UseResponseCompression();
-         app.UseMvc();
          app.UseFileServer();
          app.Run(async (context) =>
          {
-            using(var reader = new StreamReader(File.OpenRead("wwwroot/index.html")))
+            using (var reader = new StreamReader(File.OpenRead("wwwroot/index.html")))
                await context.Response.WriteAsync(reader.ReadToEnd());
          });
       }
